refacor assets state
