+++ src/redux/selectors/getPortfolioAssets.ts
@@ -4,7 +4,7 @@ import Decimal from "decimal.js";
 import { RootState } from "../store";
 import { emptySuppliedAsset, emptyBorrowedAsset, hasAssets, getRewards, toUsd } from "../utils";
 import { shrinkToken, expandToken } from "../../store";
-import { Asset, Assets } from "../assetsSlice";
+import { Asset, Assets } from "../assetState";
 import { Farm } from "../accountState";
 
 export const getPortfolioRewards = (
